requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
h11==0.14.0
idna==3.6
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
